Add presenter for angular things
