+++ pubspec.yaml
@@ -9,10 +9,11 @@ environment:
   sdk: '>=1.0.0 <2.0.0'
 
 dependencies:
+  angular: '^1.1.2+2'
   bootstrap: "^3.3.4"
   browser: '>=0.10.0 <0.11.0'
   dart_to_js_script_rewriter: '^0.1.0'
-  angular: '^1.1.2+2'
+  markdown: '^0.9.0'
   web_components: '^0.12.0+3'
 
 dev_dependencies:
