+++ web/styles/main.css
@@ -6,6 +6,10 @@ body {
     width: 100%;
 }
 
+.message-container img {
+  width: 100%;
+}
+
 .sidebar {
     position: fixed;
     top: 51px;
@@ -34,4 +38,3 @@ body {
     padding-bottom: 10px;
     border-top: 1px solid #EEEEEE;
 }
-
